Partial implementation of view.
Write code to view the practice text area and the Go! button. Write code
to dynamically generate the candidate passwords. Write code to handle
changing the number of generated passwords.

To my very pleasant surprise, when I change the number of generated
passwords, the dynamically generated (empty) list of candidate
passwords. Although it is great to see that dynamism, I really do not
want it to change until I press "Go!" I'll need to make this change.